+++ src/main/java/com/sdemo1/dto/seat/SeatDto.java
@@ -1,15 +1,13 @@
 package com.sdemo1.dto.seat;
 
-import java.math.BigInteger;
-
 /**
  * 좌석 정보 DTO
  * Java 14+ Record 클래스 사용으로 불변 객체 생성
  */
 public record SeatDto(
-    BigInteger id,
-    BigInteger concertId,
-    BigInteger seatGradeId,
+    Long id,
+    Long concertId,
+    Long seatGradeId,
     String seatRow,      // A, B, C 등 행 번호
     String seatNumber,   // 1, 2, 3 등 좌석 번호
     Integer positionX,   // 좌석의 X 좌표
